Fix Linux-only builds: uinput Pinch defense arm, i64 sensor span
- linux_uinput send gains the Pinch/PinchEnd no-op arm (Windows arm added earlier; the Linux match is a separate function).

- MtPinchTap sensor span is i64 before dividing by the i64 detent scale.

+++ kvm-platform/src/mt_pinch.rs
@@ -287,7 +287,7 @@ impl MtPinchTap {
         {
             return None;
         }
-        let span = abs_maximum(&file, ABS_MT_POSITION_X).unwrap_or(3072);
+        let span = i64::from(abs_maximum(&file, ABS_MT_POSITION_X).unwrap_or(3072));
         tracing::debug!(device = %path.display(), span, "multitouch tap node opened read-only");
         Some(TapDevice {
             file,
